alipay

+++ src/main/java/com/afd/model/order/OrderLog.java
@@ -6,7 +6,7 @@ import java.util.Date;
 public class OrderLog implements Serializable{
     private Long orderLogId;
 
-    private Integer orderId;
+    private Long orderId;
 
     private String orderCode;
 
@@ -30,11 +30,11 @@ public class OrderLog implements Serializable{
         this.orderLogId = orderLogId;
     }
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
